requirements.txt
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
deprecation==2.1.0
exceptiongroup==1.3.0
eyeD3==0.9.8
ffmpeg-python==0.2.0
filetype==1.2.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mutagen==1.47.0
packaging==25.0
pillow==11.2.1
python-telegram-bot==22.1
pytube==15.0.0
requests==2.32.4
sniffio==1.3.1
typing_extensions==4.14.0
urllib3==2.5.0
yt-dlp==2025.6.9